+++ src/models/logEntry.model.ts
