+++ src/models/posts.model.ts
