+++ src/routes/users.routes.ts
@@ -1,7 +1,7 @@
 import express, { Router } from 'express'
 import { jwtValidator } from '../middleware/jwtValidator'
-import { getUserHandler } from '../handlers/user'
-import { getPostsFromUserHandler } from '../handlers/post'
+import { getUserHandler } from '../handlers/user.handler'
+import { getPostsFromUserHandler } from '../handlers/post.handler'
 
 export const usersRouter: Router = express.Router()
